# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stockopedia Service API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "talansk@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/subscription_service/"
        }
    ],
    "paths": {
        "/subscriptions": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Get list of subscriptions",
                "description": "Return a List of subscriptions",
                "operationId": "getList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Create new subscription",
                "description": "Create new subscription",
                "operationId": "create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/subscriptions/{subscriptionId}": {
            "get": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Get subscription by id",
                "description": "Return a subscription",
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "subscriptionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Delete subscription",
                "description": "Delete subscription",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "subscriptionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Update subscription",
                "description": "Update subscription and return data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "subscriptionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SubscriptionRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "plan_code": {
                        "type": "string"
                    },
                    "monthly_cost": {
                        "type": "integer"
                    },
                    "annual_cost": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Subscription",
            "description": "API Endpoints of Project"
        }
    ]
}